Synchronize access to Schematron.pipelineSteps
Goetz 2006, p. 28: "It is a common mistake to assume that
sychronization needs to be used only when writing to shared variables;
this is simply not true. For each mutable state variable that may be
accessed by more than one thread, all accesses to that variable must
be performed with the same lock held."

Goetz, Brian. Java Concurrency in Practice. Upper Saddle River, NJ:
Addison-Wesley, 2006.